requirements.txt
fastapi==0.116.1
opencv_contrib_python==4.12.0.88
opencv_python==4.12.0.88
pydantic==2.11.7
python-dotenv==1.1.1
redis==6.2.0
Requests==2.32.4
ultralytics==8.3.169
uvicorn==0.35.0
